refactor: extract @ExceptionHandler methods into GlobalExceptionHandler
각 컨트롤러에 중복된 @ExceptionHandler를 @RestControllerAdvice로 통합하고,
WishController에 누락된 NoSuchElementException 핸들러를 추가한다.
