+++ src/main/java/shopping/config/GlobalExceptionHandler.java
@@ -1,0 +1,37 @@
+package shopping.config;
+
+import shopping.product.domain.ProfanityCheckException;
+
+import java.util.Map;
+import java.util.NoSuchElementException;
+
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.RestControllerAdvice;
+
+@RestControllerAdvice
+public class GlobalExceptionHandler {
+
+    @ExceptionHandler(IllegalArgumentException.class)
+    public ResponseEntity<Map<String, String>> handleIllegalArgument(IllegalArgumentException e) {
+        return ResponseEntity.badRequest().body(Map.of("message", e.getMessage()));
+    }
+
+    @ExceptionHandler(NoSuchElementException.class)
+    public ResponseEntity<Void> handleNotFound(NoSuchElementException e) {
+        return ResponseEntity.notFound().build();
+    }
+
+    @ExceptionHandler(ProfanityCheckException.class)
+    public ResponseEntity<Map<String, String>> handleProfanityCheckFailure(
+            ProfanityCheckException e) {
+        return ResponseEntity.internalServerError().body(Map.of("message", e.getMessage()));
+    }
+
+    @ExceptionHandler(DataIntegrityViolationException.class)
+    public ResponseEntity<Map<String, String>> handleDataIntegrityViolation(
+            DataIntegrityViolationException e) {
+        return ResponseEntity.badRequest().body(Map.of("message", "이미 존재하는 이메일입니다."));
+    }
+}
