+++ src/main/java/shopping/member/controller/MemberController.java
@@ -3,14 +3,10 @@ package shopping.member.controller;
 import shopping.member.domain.LoginMember;
 import shopping.member.domain.RegisterMember;
 
-import java.util.Map;
-
 import jakarta.validation.Valid;
 
-import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -43,14 +39,4 @@ public class MemberController {
         return ResponseEntity.ok(new MemberResponse(token));
     }
 
-    @ExceptionHandler(IllegalArgumentException.class)
-    public ResponseEntity<Map<String, String>> handleIllegalArgument(IllegalArgumentException e) {
-        return ResponseEntity.badRequest().body(Map.of("message", e.getMessage()));
-    }
-
-    @ExceptionHandler(DataIntegrityViolationException.class)
-    public ResponseEntity<Map<String, String>> handleDataIntegrityViolation(
-            DataIntegrityViolationException e) {
-        return ResponseEntity.badRequest().body(Map.of("message", "이미 존재하는 이메일입니다."));
-    }
 }
